refactor(stremio): move StremioCleanupService out of internal/api
The cleanup service has no HTTP dependencies and does not belong in
the api package. Move it to a new internal/stremio package so that
internal/api remains focused on HTTP handlers and routing.
