+++ internal/stremio/cleanup.go
@@ -1,4 +1,4 @@
-package api
+package stremio
 
 import (
 	"context"
